Add StopMode to ServerConfig
The StopMode of a server is set in the ServerConfig before the start of
the server, so that we can check if a given server supports the selected
stop mode and fail fast if it doesn't.

A new server contract ServerStopContract describes the expected behavior
of the various stop modes.

Immediate (default):
  Has to be supported by all servers, quickest way to stop a server

Graceful with timeout:
  Will wait for inflight requests and stop as soon as possible

Delayed:
  Used for servers that don't support Graceful stop mode but will allow
  inflight requests to continue for a fixed timeout duration